LP-1246 Moved capital contribution validator into validator folder
